Improve certificate handling
The mail server will be using TLS 1.2 rather than SSL so the SSL setting has been removed and replaced with a flag that indicates whether to allow an invalid certificate (for use in test evironments). The setting will inherit from the corresponding app level flag but can also be overridden in the SMTP settings.

AB#5208